Use NTP puppet module from forge.puppetlabs.com
After some discussion in #openstack-infra using modules from puppetforge
is a good thing.  So, this is a simple change to use an existing module
that I maintain.

It is also a good starting point for me to get more involved with
openstack CI team.

Additionally, I've updated some of the documentation around the
installation of a puppetmaster.

Change-Id: I95ace224ba489f913daa61efd10021e7a0274970
Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>